Clarify the handler result type to be "any"
Previously, `Result(object)` was expected from the handlers despite `Any` was essentially meant for the public protocols. `Result` is only needed for internal type-checking — to avoid using any other value instead of the results by accident. Now, `Optional[object]` is specified for the protocols as seen by the users, and it is casted to `Optional[Result]` where needed.

There is no need to declare the protocol results as `Any` because any actual type or structure becomes a special case of `Optional[object]` anyway. The latter form allows us to be strict for not using any methods/fields/properties and to make no implicit assumptions on the result type.

Signed-off-by: Sergey Vasilyev <nolar@nolar.info>