Remove leading zeros from IPv4 in tests
Leading zeros are no longer tolerated since Python 3.10 and 3.9.5 (see https://docs.python.org/3/library/ipaddress.html#ipaddress.IPv4Address). This was breaking the tests due to a mismatch of the IP addresses in SSL certificates, as they were treated as regular hostnames after this change, and includes "as is" (not compacted).

Signed-off-by: Sergey Vasilyev <nolar@nolar.info>