Rename internal entities to not mention "resource(s)" explicitly
Everything in the Kubernetes operators is about resources anyway — there is no need to highlight it. Originally, this "resource" prefix was added to distinguish from "activities". Over time, it became clear that "activities" is the only non-resource entity.

This change is intended to simplify the codebase and clean up the code visually.

Signed-off-by: Sergey Vasilyev <nolar@nolar.info>