Fix an issue with Click 8.0.0 for custom type double-conversion
Presumably caused by this change in Click:

* https://github.com/pallets/click/commit/e79c2b47aa5b456e6c70f980ecf0f883447eb340

> … This ensures the default value is processed like other values. Some type converters were adjusted to accept values that are already the correct type. …

Previously, the default string was converted to the default value of the parameter only once. With Click 8.0.0, first, the default string ("full") is converted to its typed (enum), and then its enum is converted again as a regular value.

According to the docs (from a quote below), this is a violation of the conversion contract from Click's side — i.e. not only a string is passed:

> … To implement a custom type, you need to subclass the ParamType class. Override the convert() method to convert the value **from a string to the correct type**. … https://click.palletsprojects.com/en/7.x/parameters/#implementing-custom-types

With this fix, if a value is already of the proper type, do not convert it and use it "as is".

Signed-off-by: Sergey Vasilyev <nolar@nolar.info>