Fall back to rudimentary login handlers if the clients are absent
This will make Kopf runnable in the majority of cases even with no `pykube-ng` and `kubernetes` client libraries installed. This was not a problem till recently, but then `pykube-ng` was removed from the implicit dependencies, so users get frustrated by "unauthenticated" or "ran out of credentials" errors.

To avoid running too many login handlers at startup, the rudimentary login handlers are added strictly when the client libraries are absent, not always. This can be overridden with custom login handlers calling these functions (as documented).

Signed-off-by: Sergey Vasilyev <nolar@nolar.info>