Simplify signatures of the login functions
Positional args are not used, and positional args are not a protocol for handlers anymore. They were originally added to match the callback signatures when invocation had positional args (when filters/lifecycles were called via the invocation routines) — but this is not the case anymore. So, remove them. Keep only the kwargs.

Signed-off-by: Sergey Vasilyev <nolar@nolar.info>