Lint the source code with pre-commit
The time has come: there are 3+ linters & checkers in the CI and in local development — mypy, isort (code & examples separately), lint-imports — which are easier to execute at once with pre-commit.

Tests are not included in the pre-commit hooks — they take quite some time (2 mins), which can be too slow for every commit and even for every push. This can be reconsidered later.

Signed-off-by: Sergey Vasilyev <nolar@nolar.info>