Skip namespaces for cluster-scoped resources/peerings in tests
Some tests were improperly designed to always use a hard-coded namespace (e.g. "ns1") — even if a cluster-scoped resource was used.

This, in theory, should break `resource.get_url()` calls, as it is an error to provide a specific namespace to a non-namespaced (cluster-scoped) resource.

However, it was never noticed because almost all `resource.get_url()` calls happen inside of the API client functions which are/were mocked in all tests — so the URL construction did not happen. (A permanent reminder: mocks a bad!)

Only when the lower-level API routines were extracted and mocked (for specific HTTP methods: GET/POST/PATCH/DELETE/etc), and the higher-level routines stopped being mocked, it has become noticed that some tests attempt to feed a namespace to cluster-scoped resources.

This change fixes the tests (not the code!) to provide a namespace that corresponds to the currently tested resource (specifically, its scope).

As for the code fix — this is actually a minor bug visible only in tests: when a cluster-scoped peering resource is used in a namespaced setup. This does not happen in real operators too. But it seems appropriate not to start multiple keep-alive tasks for a single peering resource (one per namespace) — exactly as it is already done for regular resources and watchers several lines below.

Signed-off-by: Sergey Vasilyev <nolar@nolar.info>