Redesign a test with namespaced peering in cluster-only case
It happened so that a namespaced peering resource was used in a test which implies only the cluster peering resource and cluster processes. If such case would happen in a real operator, it would watch the events from all namespaces peering in the cluster. However, this does not happen in real operators because a proper peering resource is always selected for the operator's scope — so it makes no sense to test unrealistic (and misbehaving) case.

The "insights" fixture is adjusted to populate the backbone only with peering resources actually used in every test, and to not bring the peering resource as a fixture on its own (otherwise, it would cross-join all "peering_resource" values with tests' "cluster_peering_resource" values, thus breaking the setup).

Signed-off-by: Sergey Vasilyev <nolar@nolar.info>