Return the connection to the pool as soon as parsed
By default, this was done by `__del__` destructor once the `response` object was not referred anymore — i.e. immediately at the end of the get/post/patch/delete routines. So, there was no actual leak or a bug (except if the garbage collector is disabled). Now, the responses & connections will be released to the pool and cleaned up explicitly, even with the garbage collector disabled.

This change does not apply to routines that get the raw response object for their own processing — e.g. streaming. Such routines do their own part in releasing the system resources properly. This change is only for those routines where we discard the response objects ourselves.

Signed-off-by: Sergey Vasilyev <nolar@nolar.info>