Exclude admission-only resources from streaming/watching
They were added by mistake: if there is an admission handler for a resource and no other handlers except admission, there is no need to start watching for this resource and for its updates. Among other things, the operator might be lacking RBAC permissions to do so — since it does not operate that resource. However, if there are admission handlers and some other handlers, the streaming/watching will happen as usual.

Introduced in 9c3bd1a378835ef2e65051ea2b21e6ab42a2c7ce by cloning the behaviour of other handler types without considering the actual need (or the lack of it).

Signed-off-by: Sergey Vasilyev <nolar@nolar.info>