Split the resources into specialised set and stop misusing them
Split the all-in-one set of resources into 3 separate sets: "watched", "indexed", and "webhook" — to be used in separate engines or for separate purposes. Unless there is an overlap via handlers, the resource sets will not be overlapping anymore.

As a result, the webhook-only resources will not be watched anymore, i.e. no watching tasks will be spawned (which required RBAC permissions and failed). The opposite is also true — the watched-only resources will not be considered in admissions — but it was already avoided by checking for webhook handlers even if all resources were scanned there.

Signed-off-by: Sergey Vasilyev <nolar@nolar.info>