Replace aiojobs with a simplified "fire-and-forget" scheduler
Python 3.10 removed some features explicitly used in aiojobs, such as `loop=` kwargs and `asyncio.get_event_loop()` (see code comments for details). Unfortunately, aiojobs seems essentially unmaintained for 2 years (since July 2019), so there is no hope it will be fixed soon. To keep Kopf compatible with Python 3.10, aiojobs' scheduler is replaced with a self-made and over-simplified "fire-and-forget" task scheduler — so that aiojobs can be removed as a dependency.

Note: enabling Python 3.10 is done as a separate change. This commit is only for refactoring.

Signed-off-by: Sergey Vasilyev <nolar@nolar.info>