Stop daemons on operator exit
The issue was introduced with a self-made "fire-and-forget" scheduler, specifically with its closing logic: the central daemon killer was spawning the daemon-stopping tasks for each daemon on operator exit, but every such task was immediately cancelled because the scheduler was closed and awaited in `.close()`.

To fix that, it now gracefully awaits for all the fresh daemon-stopping tasks to finish before closing the scheduler (which will have nothing to cancel except for its own supporting tasks).

Signed-off-by: Sergey Vasilyev <nolar@nolar.info>