Add old, new, diff and operation to WebhookCause
Adding kwargs old, new, diff and operation to the WebhookCause and
passing them to the appropriate handler. This allows users to write
more powerful webhooks and allows them to create validation and
mutation webhooks that can be look at the old object to determine
their behavior. This commit also adds the operation field to the
passed kwargs to allow users to create more powerful/flexibly wenhooks.

Signed-off-by: Sambhav Kothari <sambhavs.email@gmail.com>