Cleanup the system resources of webhook servers/tunnels
The webhook server/tunnel protocol is extended with an optional part: it can now be an asynchronous context manager. This is a more explicit way of cleaning up the system resources than relying on the `finally:` block or the `GeneratorExit` exception. The latter is not guaranteed to happen until the generator is garbage-collected, which can happen much later than expected (or never). In that case, the resources remain leaked and can fail the tests. For example, this was happening in PyPy 3.8.

The webhook servers & tunnels provided by Kopf support both protocols: the simple callable protocol (with the cleanup on garbage collection in the `finally:` block) and the async context manager (with the explicit cleanup on exiting).

Signed-off-by: Sergey Vasilyev <nolar@nolar.info>