Restrict the individual test execution time
Normally, individual tests should not take longer than 1 second, preferably closer to 0 seconds. Some selected tests, however, are slow by their nature (e.g. ngrok or kopf-runner) — so they take around 1-2 seconds. But this limit is only applicable to CI — not to local dev-mode runs, where breakpoints and infinite pauses are needed.

This change should prevent the test-runs from being stuck at one single test if the test is designed improperly or because of the environment issues. This removes the need to keep the timing safeguards inside the tests, which complicates the tests themselves (e.g. with `async-timeout`) — unless such safe-guards are a part of the test design.

Signed-off-by: Sergey Vasilyev <nolar@nolar.info>