invocation: Warn when no more free workers
Sync handlers and daemons are run inside a thread for each resource.
When the number of resources exceeds the default number of allocated
workers (for example `os.cpu_count() + 4` in python 3.9), then they are
kept pending and will thus never be launched.

As explained in the documentation, this number can be tweaked. On top of
that, this commit makes a warning appear if the limit is exceeded.

Signed-off-by: Frank Villaro-Dixon <frank.villaro@infomaniak.com>