Define credentials expiration time (optionally)
The re-authentication happens on the same trigger as before: when no valid credentials are left in the vault when they are needed.

The same as before, there is no preliminary vault population or credentials refreshment, including on credentials expiration: although we know that the credentials have expired, the re-authentication does not happen until an API call is attempted.

The existing API calls initiated before the expiration —e.g. the watch-streams— will continue running and will not be interrupted until closed from the server side or timed out due to settings.

Signed-off-by: Sergey Vasilyev <nolar@nolar.info>