Fix the asyncio guards to expect changes during the task lising
The WeakSet can change while it is iterated and being converted to a list (this seems fast, but it is not always the case and tends to be reproducible in the same setup). This fix guards against that and tries to convert the WeakSet to a list 1000 times before giving up — the same as Python's asyncio does.

Signed-off-by: Sergey Vasilyev <nolar@nolar.info>