Utilise uvloop automatically if present for CLI & synchronous `run()`
`uvloop` is a drop-in replacement for asyncio's event loops, which promises 2x–2.5x faster execution of the same i/o code (http://magic.io/blog/uvloop-blazing-fast-python-networking/). There is no reason to ignore it if it is installed. For convenience, Kopf can be installed as `pip install kopf[uvloop]` to bring the mode as a dependency.

The implicit activation only affects the CLI mode — i.e. `kopf run …`. To disable, uninstall `uvloop`.

For in-code invocations —such as `kopf.run()` or `KopfRunner`— it will use the preinstalled event loop policy (or the default one). `uvloop` is not activated implicitly if Kopf is embedded into a custom application.

Signed-off-by: Sergey Vasilyev <nolar@nolar.info>