Get rid of async-level low-level debug logs
The debug mode was needed at the earlier stages of the development but is not needed anymore. However, it brings much inconvenience:

The event loop must be known already at the stage of logging configuration, while it is normally created much later in the `kopf.run()`. The `get_event_loop()`, in turn, emits a warning about a deprecated use with no event loop set in advance, which is converted to errors in tests. It seems, the event loop management is supposed to be explicit — so be it.

However, trying to `set_event_loop()` up in the stack if CLI commands brings other issues and warnings: there is no place to close that newly created loop unless decorators with context managers are added for all CLI commands. This seems an overkill for the unneeded debug mode. There is the "proper" way to run an implicit loop — `asyncio.run()` — so let's use it.

Signed-off-by: Sergey Vasilyev <nolar@nolar.info>