Migrate all API tests from aresponses to KMock
Kmock is a Kubernetes Mock Server, but can mock any API and HTTP servers locally. It allows expressing any regular or weird (mis)behaviour of the server side as needed for local tests — and is much shorter and compact than the equivalent code with `aresponses`.

* https://kmock.readthedocs.io/
* https://github.com/nolar/kmock

It originally started as a couple of helper tools inside Kopf itself, but I later extracted it into a separate library.

Signed-off-by: Sergey Vasilyev <nolar@nolar.info>