Ensure the reason is never an empty string, but None in HTTP responses
Otherwise, `aiohttp` (in tests) cannot parse the response with the trailing space in the status — `"HTTP/1.1 555 "` — since `"555 ".split()` returns only one element, and the whole string `"555 "` is used as the status with expectations of being a 3-digit numeric.

For some reason, the tests fail with this only in PyPy (all versions).

Signed-off-by: Sergey Vasilyev <nolar@nolar.info>