Separate --dev from --priority to prevent overriding the env var
If `--dev` is stored as a specific value of another variable (`priority`), and that variable has he env var defined (`KOPF_RUN_PRIORITY`), that value is lost, and the default value of `dev=None` overrides it.

To solve it, separate `--dev` into its own variable and set the specific value directly in code.

Signed-off-by: Sergey Vasilyev <nolar@nolar.info>