Migrate from setup.py to pyproject.toml, but keep setuptools
The approach with executable `setup.py` is considered a legacy and insecure — it allows arbitrary code execution on the developers & building machines if a package repo is mistyped or only tried out. Developers might enable the safe mode where only the non-executable dependencies are installed, i.e. those based on PEP-621 (`pyproject.toml`).

Setuptools and pip are a very lightweight and pre-installed toolkit for packaging, which is more than sufficient for Kopf. As an extra, they allow using `setuptools-scm` and continue using git tags for versioning, which in turn enabled the GitHub UI for Releases.

An alternative: Poetry — but it is heavy-weight and brings a lot of dependencies, which might conflict with the project itself, but it does not bring any benefits compared to setuptools. It might be good for big projects & apps, but seems like an overkill for a tiny library. Especially since some dependencies are optimised for the disk size and extracted into extras for this reason.

Signed-off-by: Sergey Vasilyev <nolar@nolar.info>