Switch the deprecated routines for Python 3.14 (asyncio.iscoroutinefunction)
The legacy `asyncio.iscoroutinefunction()` is deprecated and issues a warning. In strict mode, the warning raises the error and fails all the server-side processing, thus also failing all the API tests.

We can remove it from Kopf's code, including tests. But it remains in aresponses — so, suppress that warning explicitly.

This change is only relevant for tests, not for the framework itself — that was fixed earlier.

Signed-off-by: Sergey Vasilyev <nolar@nolar.info>